Telemetry: review comments; move more stuff to separate thread
