+++ dd-java-agent/appsec/src/main/java/com/datadog/appsec/dependency/DependencyServiceImpl.java
@@ -5,8 +5,13 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.lang.instrument.Instrumentation;
+import java.lang.invoke.MethodHandle;
+import java.lang.invoke.MethodHandles;
+import java.lang.invoke.MethodType;
+import java.lang.reflect.UndeclaredThrowableException;
 import java.net.JarURLConnection;
 import java.net.URI;
+import java.net.URISyntaxException;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -28,7 +33,8 @@ public class DependencyServiceImpl implements DependencyService {
 
   private static final String JAR_SUFFIX = ".jar";
 
-  private Set<URI> currentSet; // guarded by this
+  private Set<URL> currentSet; // guarded by this
+  private volatile JbossVirtualFileHelper jbossVirtualFileHelper;
 
   public DependencyServiceImpl() {
     this.currentSet = new HashSet<>();
@@ -45,17 +51,22 @@ public class DependencyServiceImpl implements DependencyService {
 
   @Override
   public Collection<Dependency> determineNewDependencies() {
-    Set<URI> currentSet;
+    Set<URL> currentSet;
     synchronized (this) {
       currentSet = this.currentSet;
       this.currentSet = new HashSet<>();
     }
 
     List<Dependency> deps = new ArrayList<>(currentSet.size());
-    for (URI uri : currentSet) {
+    for (URL url : currentSet) {
       if (Thread.interrupted()) {
         log.warn("Interrupted while processing dependencies");
         break;
+      }
+
+      URI uri = convertToURI(url);
+      if (uri == null) {
+        continue;
       }
 
       Dependency dep = identifyLibrary(uri);
@@ -77,16 +88,41 @@ public class DependencyServiceImpl implements DependencyService {
   }
 
   @Override
-  public void addURI(URI uri) {
+  public void addURL(URL url) {
     // we ignore .class files directly within webapp folder (they aren't part of dependencies)
-    String path = uri.getPath();
+    String path = url.getPath();
     if (path != null && path.endsWith(".class")) {
       return;
     }
 
     synchronized (this) {
-      currentSet.add(uri);
+      currentSet.add(url);
     }
+  }
+
+  private URI convertToURI(URL location) {
+    URI uri = null;
+
+    if (location.getProtocol().equals("vfs")) {
+      // resolve jboss virtual file system
+      try {
+        uri = getJbossVfsPath(location);
+      } catch (RuntimeException rte) {
+        log.debug("Error in call to getJbossVfsPath", rte);
+        return null;
+      }
+    }
+
+    if (uri == null) {
+      try {
+        uri = location.toURI();
+      } catch (URISyntaxException e) {
+        log.warn("Error converting URL to URI", e);
+        // silently ignored
+      }
+    }
+
+    return uri;
   }
 
   /**
@@ -131,7 +167,7 @@ public class DependencyServiceImpl implements DependencyService {
     }
 
     Dependency dependency = null;
-    try (JarFile file = new JarFile(jar);
+    try (JarFile file = new JarFile(jar, false /* no verify */);
         InputStream is = new FileInputStream(jar)) {
 
       // Try to get from maven properties
@@ -198,6 +234,91 @@ public class DependencyServiceImpl implements DependencyService {
       return file.getManifest();
     } catch (IOException e) {
       // silently ignored
+    }
+    return null;
+  }
+
+  static class JbossVirtualFileHelper {
+    private final MethodHandle getPhysicalFile;
+    private final MethodHandle getName;
+
+    public static final JbossVirtualFileHelper FAILED_HELPER = new JbossVirtualFileHelper();
+
+    public JbossVirtualFileHelper(ClassLoader loader) throws Exception {
+      MethodHandles.Lookup lookup = MethodHandles.lookup();
+      Class<?> virtualFileCls = loader.loadClass("org.jboss.vfs.VirtualFile");
+      getPhysicalFile =
+          lookup.findVirtual(virtualFileCls, "getPhysicalFile", MethodType.methodType(File.class));
+      getName = lookup.findVirtual(virtualFileCls, "getName", MethodType.methodType(String.class));
+    }
+
+    private JbossVirtualFileHelper() {
+      getPhysicalFile = null;
+      getName = null;
+    }
+
+    public File getPhysicalFile(Object virtualFile) {
+      try {
+        return (File) getPhysicalFile.invoke(virtualFile);
+      } catch (Throwable e) {
+        throw new UndeclaredThrowableException(e);
+      }
+    }
+
+    public String getName(Object virtualFile) {
+      try {
+        return (String) getName.invoke(virtualFile);
+      } catch (Throwable e) {
+        throw new UndeclaredThrowableException(e);
+      }
+    }
+  }
+
+  private URI getJbossVfsPath(URL location) {
+    JbossVirtualFileHelper jbossVirtualFileHelper = this.jbossVirtualFileHelper;
+    if (jbossVirtualFileHelper == JbossVirtualFileHelper.FAILED_HELPER) {
+      return null;
+    }
+
+    Object virtualFile;
+    try {
+      virtualFile = location.openConnection().getContent();
+    } catch (IOException e) {
+      // silently ignored
+      return null;
+    }
+    if (virtualFile == null) {
+      return null;
+    }
+
+    if (jbossVirtualFileHelper == null) {
+      try {
+        jbossVirtualFileHelper =
+            this.jbossVirtualFileHelper =
+                new JbossVirtualFileHelper(virtualFile.getClass().getClassLoader());
+      } catch (Exception e) {
+        log.warn("Error preparing for inspection of jboss virtual files", e);
+        return null;
+      }
+    }
+
+    // call VirtualFile.getPhysicalFile
+    File physicalFile = jbossVirtualFileHelper.getPhysicalFile(virtualFile);
+    if (physicalFile.isFile() && physicalFile.getName().endsWith(".jar")) {
+      return physicalFile.toURI();
+    } else {
+      log.info("Physical file {} is not a jar", physicalFile);
+    }
+
+    // not sure what this is about, but it's what the old code used to do
+    // this is not correct as a general matter, since getName returns the virtual name,
+    // which may not match the physical name
+    // The original comment reads:
+    // "physical file returns 'content' folder, we manually resolve to the actual jar location"
+    String fileName = jbossVirtualFileHelper.getName(virtualFile);
+    physicalFile = new File(physicalFile.getParentFile(), fileName);
+    if (physicalFile.isFile()) {
+      return physicalFile.toURI();
     }
     return null;
   }
