+++ telemetry/src/main/java/datadog/telemetry/TelemetryServiceImpl.java
@@ -11,7 +11,6 @@ import datadog.telemetry.api.Metric;
 import datadog.telemetry.api.Payload;
 import datadog.telemetry.api.RequestType;
 import datadog.trace.api.time.TimeSource;
-
 import java.util.*;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
@@ -64,11 +63,7 @@ public class TelemetryServiceImpl implements TelemetryService {
   @Override
   public boolean addConfiguration(Map<String, Object> configuration) {
     for (Map.Entry<String, Object> entry : configuration.entrySet()) {
-      if (!this.configurations.offer(
-          new KeyValue()
-              .name(entry.getKey())
-              .value(entry.getValue())
-      )) {
+      if (!this.configurations.offer(new KeyValue().name(entry.getKey()).value(entry.getValue()))) {
         return false;
       }
     }
